# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/knowledge/upload": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload Files",
        "description": "上传文件到知识库——异步清洗、分块、向量化、存储。\n\n支持 PDF/DOCX/XLSX/CSV/HTML/MD/TXT 及代码文件。\n图片和二进制文件会被跳过并返回警告。",
        "operationId": "upload_files_api_v1_projects__project_id__knowledge_upload_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_api_v1_projects__project_id__knowledge_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestResponse"
                  },
                  "title": "Response Upload Files Api V1 Projects  Project Id  Knowledge Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/knowledge/files": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Files",
        "description": "列出项目已索引文件。",
        "operationId": "list_files_api_v1_projects__project_id__knowledge_files_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileInfo"
                  },
                  "title": "Response List Files Api V1 Projects  Project Id  Knowledge Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/knowledge/files/{file_id}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete File",
        "description": "删除文件及其所有分块和向量数据。",
        "operationId": "delete_file_api_v1_projects__project_id__knowledge_files__file_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/knowledge/search/quick": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Quick Search",
        "description": "快速模糊搜索——纯 pg_trgm，毫秒级，供知识库面板使用。\n\n同时匹配文档内容和文件名，按相似度排序。\n不做向量检索、Query 扩写或 LLM 重排。",
        "operationId": "quick_search_api_v1_projects__project_id__knowledge_search_quick_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuzzySearchItem"
                  },
                  "title": "Response Quick Search Api V1 Projects  Project Id  Knowledge Search Quick Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Knowledge",
        "description": "检索知识库——Query 扩写 + 混合检索 + RRF 融合 + LLM 重排。",
        "operationId": "search_knowledge_api_v1_projects__project_id__knowledge_search_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_files_api_v1_projects__project_id__knowledge_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_files_api_v1_projects__project_id__knowledge_upload_post"
      },
...
      "DeleteResponse": {
        "properties": {
          "deleted_chunks": {
            "type": "integer",
            "title": "Deleted Chunks"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "deleted_chunks",
          "message"
        ],
        "title": "DeleteResponse",
        "description": "删除响应。"
      },
...
      "FileInfo": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "ingested_at": {
            "type": "string",
            "title": "Ingested At"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "file_name",
          "file_type",
          "chunk_count",
          "ingested_at"
        ],
        "title": "FileInfo",
        "description": "已索引文件信息。"
      },
      "FuzzySearchItem": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "file_name",
          "file_type",
          "file_id",
          "chunk_index",
          "content",
          "score"
        ],
        "title": "FuzzySearchItem",
        "description": "轻量模糊搜索结果项——文件名匹配优先。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "chunks_created": {
            "type": "integer",
            "title": "Chunks Created"
          },
          "chunks_skipped": {
            "type": "integer",
            "title": "Chunks Skipped"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "file_id",
          "chunks_created",
          "chunks_skipped"
        ],
        "title": "IngestResponse",
        "description": "摄入结果响应。"
      },
...
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Results"
          },
          "expanded_queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expanded Queries"
          },
          "used_expansion": {
            "type": "boolean",
            "title": "Used Expansion"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "results",
          "expanded_queries",
          "used_expansion",
          "total"
        ],
        "title": "SearchResponse",
        "description": "检索响应。"
      },
      "SearchResultItem": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "file_name",
          "file_type",
          "file_id",
          "chunk_index",
          "content",
          "score"
        ],
        "title": "SearchResultItem",
        "description": "检索结果项。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}